Fix MSVC compile error
Fixes #296

+++ cpp/neuralnet/eigenbackend.cpp
@@ -365,7 +365,7 @@ struct ConvLayer {
       TENSOR4 kernel = TensorMap<const Tensor<const SCALAR, 4>>(desc.weights.data(), convXSize, convYSize, inChannels, outChannels);
       imagePatchSize = convXSize * convYSize * inChannels;
       Eigen::array<int, 4> dimensionPermutatation({3, 2, 0, 1});
-      Eigen::array<int, 2> newShape({outChannels, imagePatchSize});
+      Eigen::array<int64_t, 2> newShape({outChannels, imagePatchSize});
       imagePatchKernel = kernel.shuffle(dimensionPermutatation).reshape(newShape);
     }
   }
